+++ firm/include/wifi_plsi.h
@@ -5,9 +5,9 @@
 // Globals variables for WiFi task
 //--------------------------------------------------------------------------------
 
-extern uint8_t configFirmwareEnabled;
-extern uint8_t configFirmwareSequence;
-extern uint8_t configFirmwareMemory;
+extern uint16_t configFirmwareEnabled;
+extern uint16_t configFirmwareSequence;
+extern uint16_t configFirmwareMemory;
 
 //--------------------------------------------------------------------------------
 // Wifi task main functions
